apps/worker/requirements.txt
celery==5.4.0
redis==5.2.1
pandas==2.2.3
openpyxl==3.1.5
boto3==1.35.81
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
python-dotenv==1.0.1
scipy==1.14.1
reportlab==4.2.5
matplotlib==3.9.4
python-dateutil==2.9.0
